add changes and FIXME notes to make compatible with API

+++ beatrix/src/test/java/org/killbill/billing/beatrix/integration/usage/TestUsageInArrear.java
@@ -587,23 +587,23 @@ public class TestUsageInArrear extends TestIntegrationBase {
         subscriptionChecker.checkSubscriptionCreated(bp2.getId(), internalCallContext);
 
         final List<UsageRecord> bp1StoneRecords1 = new ArrayList();
-        bp1StoneRecords1.add(new UsageRecord(new LocalDate(2012, 4, 5), BigDecimal.valueOf(5L)));
-        bp1StoneRecords1.add(new UsageRecord(new LocalDate(2012, 4, 15), BigDecimal.valueOf(10L)));
-        bp1StoneRecords1.add(new UsageRecord(new LocalDate(2012, 4, 16), BigDecimal.valueOf(15L)));
+        bp1StoneRecords1.add(new UsageRecord(new LocalDate(2012, 4, 5), BigDecimal.valueOf(5L).longValue() /* FIXME-1469 : API backward compat */));
+        bp1StoneRecords1.add(new UsageRecord(new LocalDate(2012, 4, 15), BigDecimal.valueOf(10L).longValue() /* FIXME-1469 : API backward compat */));
+        bp1StoneRecords1.add(new UsageRecord(new LocalDate(2012, 4, 16), BigDecimal.valueOf(15L).longValue() /* FIXME-1469 : API backward compat */));
         final SubscriptionUsageRecord bp1UsageRecord1 = new SubscriptionUsageRecord(bp1.getId(), "bp1-tracking-1", ImmutableList.of(new UnitUsageRecord("stones", bp1StoneRecords1)));
         recordUsageData(bp1UsageRecord1, callContext);
 
         final List<UsageRecord> bp1StoneRecords2 = new ArrayList();
-        bp1StoneRecords2.add(new UsageRecord(new LocalDate(2012, 4, 23), BigDecimal.valueOf(10L)));
+        bp1StoneRecords2.add(new UsageRecord(new LocalDate(2012, 4, 23), BigDecimal.valueOf(10L).longValue() /* FIXME-1469 : API backward compat */));
         // Outside of range for this period -> Its tracking ID spreads across 2 invoices
-        bp1StoneRecords2.add(new UsageRecord(new LocalDate(2012, 5, 1), BigDecimal.valueOf(101L)));
+        bp1StoneRecords2.add(new UsageRecord(new LocalDate(2012, 5, 1), BigDecimal.valueOf(101L).longValue() /* FIXME-1469 : API backward compat */));
         final SubscriptionUsageRecord bp1UsageRecord2 = new SubscriptionUsageRecord(bp1.getId(), "bp1-tracking-2", ImmutableList.of(new UnitUsageRecord("stones", bp1StoneRecords2)));
         recordUsageData(bp1UsageRecord2, callContext);
 
         final List<UsageRecord> bp2StoneRecords = new ArrayList();
-        bp2StoneRecords.add(new UsageRecord(new LocalDate(2012, 4, 5), BigDecimal.valueOf(85L)));
-        bp2StoneRecords.add(new UsageRecord(new LocalDate(2012, 4, 15), BigDecimal.valueOf(150L)));
-        bp2StoneRecords.add(new UsageRecord(new LocalDate(2012, 4, 16), BigDecimal.valueOf(39L)));
+        bp2StoneRecords.add(new UsageRecord(new LocalDate(2012, 4, 5), BigDecimal.valueOf(85L).longValue() /* FIXME-1469 : API backward compat */));
+        bp2StoneRecords.add(new UsageRecord(new LocalDate(2012, 4, 15), BigDecimal.valueOf(150L).longValue() /* FIXME-1469 : API backward compat */));
+        bp2StoneRecords.add(new UsageRecord(new LocalDate(2012, 4, 16), BigDecimal.valueOf(39L).longValue() /* FIXME-1469 : API backward compat */));
         final SubscriptionUsageRecord bp2UsageRecord = new SubscriptionUsageRecord(bp2.getId(), "bp2-tracking-1", ImmutableList.of(new UnitUsageRecord("stones", bp2StoneRecords)));
         recordUsageData(bp2UsageRecord, callContext);
