+++ jaxrs/src/main/java/org/killbill/billing/jaxrs/json/InvoiceItemJson.java
@@ -134,7 +134,7 @@ public class InvoiceItemJson extends JsonBase {
              item.getInvoiceItemType(),
              item.getDescription(), item.getStartDate(), item.getEndDate(),
              item.getAmount(), item.getRate(), item.getCurrency(),
-             item.getQuantity(), item.getItemDetails(), item.getCatalogEffectiveDate(),
+             BigDecimal.valueOf(item.getQuantity()) /* FIXME-1469 : API backward compat */, item.getItemDetails(), item.getCatalogEffectiveDate(),
              toInvoiceItemJson(childItems), toAuditLogJson(auditLogs));
     }
 
@@ -258,7 +258,7 @@ public class InvoiceItemJson extends JsonBase {
             }
 
             @Override
-            public BigDecimal getQuantity() { return quantity; }
+            public Integer getQuantity() { return quantity.intValue(); /* FIXME-1469 : API backward compat */ }
 
             @Override
             public String getItemDetails() { return itemDetails; }
